docker/4.建立fastapi和postgreSQL/requirements.txt
requests
streamlit
psycopg2-binary